requirements.txt
flask
requests
google-generativeai
Pillow